Move openidConf to middleware package
