+++ api/middleware/openidConf_test.go
@@ -1,4 +1,4 @@
-package server
+package middleware
 
 import (
 	"crypto/rsa"
