+++ api/server/http.go
@@ -8,6 +8,7 @@ import (
 	jwt "github.com/dgrijalva/jwt-go"
 	_ "github.com/equinor/oneseismic/api/docs" // docs is generated by Swag CLI, you have to import it.
 	l "github.com/equinor/oneseismic/api/logger"
+	"github.com/equinor/oneseismic/api/middleware"
 	claimsmiddleware "github.com/equinor/oneseismic/api/middleware/claims"
 	jwtmiddleware "github.com/iris-contrib/middleware/jwt"
 	prometheusmiddleware "github.com/iris-contrib/middleware/prometheus"
@@ -64,7 +65,7 @@ func Configure(hs *HTTPServer, opts ...HTTPServerOption) error {
 func WithOAuth2(oauthOpt OAuth2Option) HTTPServerOption {
 
 	return newFuncOption(func(hs *HTTPServer) error {
-		sigKeySet, err := GetOIDCKeySet(oauthOpt.AuthServer)
+		sigKeySet, err := middleware.GetOIDCKeySet(oauthOpt.AuthServer)
 		if err != nil {
 			return fmt.Errorf("Couldn't get keyset: %v", err)
 		}
